Add pipeline and fix panic handler for docs
This pipeline adds linting, building, and documentation generation.

Linting has two stages: normal and pedantic. The normal stage has the regular lints available with Clippy. The pedantic stage contains additional linting that is allowed to fail. It exists to display to the user if there are any potential issues beyond the ones that are detectable by the normal stage. However, this additional linting can have false positives, which is why it is allowed to fail.

The building stage ensures that the Rust projects are able to be built.

The documentation generation stage generates documentation (public and internal), and uploads it as an artifact to the job. However, since GitLab Pages is disabled on git.ucsc.edu, this pipeline stage will only be able to generate an artifact, instead of being displayed as webpages.

The panic handlers were preventing the documentation from being generated. This has been fixed by moving panic-semihosting to a feature of each crate.