Add EEPROM Module, secret entropy source, make tests debug only
Adds the EEPROM module, which allows reading from and writing to the EEPROM module. The EEPROM layout is defined in a separate crate to allow the build script to generate EEPROM files. Reads and writes with the EEPROM module have been tested in the `tests` crate.

Implements the secret entropy source using the EEPROM module.

Tests have been made debug-only to allow the use of debug-only functions. Use debug-only functions where needed to expose private functionality to tests.