Finish implementing the clock drift entropy source
This implementation of clock drift takes the LSB of a counter in a tight loop for the time it takes to reach 16 subseconds. A subsecond is one cycle of the internal hibernation clock, which is 32.768 KHz. Therefore, a sample is taken every 1/2048 of a second. Samples are taken over a 1/2 second period.