Add .text_jump section to work with bootloader
The bootloader simply jumps to 0x8000 to pass control to our firmware, so we need to set the starting SP, vector table, and jump to the reset handler. The starting SP and vector table is set with the `set-sp` and `set-vtor` features of `cortex-m-rt`. The reset handler jump is handled by creating a `.text_jump` section which contains a branch instruction to the reset handler. This `.text_jump` section is placed at the beginning of our flash address space.

Fixes #9.